EulerProblem4 reset to standard problem
I reset EulerProblem4 to solve the problem as described rather than
investigating the the distribution of factors of palindromes.  If you
want to see every palindromic product of a pair of n-digit factors, then
go to the last commit for this file.  As it stands, this is responsive
to problem 4 (though make sure you set range in line 18 to (1,7).